test(ui): point restored composer-lane tests at the split history modules
